refactor: move favorites module to core module
