+++ src/core/favorites/entities/favorites.entity.ts
