+++ src/core/favorites/favorites.controller.spec.ts
