+++ src/core/favorites/favorites.module.ts
