+++ src/core/favorites/favorites.service.spec.ts
